+++ config/test.exs
@@ -1,5 +1,1 @@
 import Config
-
-config :lightpanda,
-  cdp_timeout: 50,
-  binary_release: "nightly"