+++ test/test_helper.exs
@@ -1,5 +1,1 @@
 ExUnit.start()
-
-Mimic.copy(Lightpanda.CDP.Client)
-Mimic.copy(Lightpanda.HTTP)
-Mimic.copy(Lightpanda.Runner)