+++ CHANGELOG.md
@@ -5,4 +5,4 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Features
-- Initial Elixir wrapper scaffold for Lightpanda
+- Initial Rustler-backed Servox scaffold